refactor: separate socket handlers into separate files
